ai-resume-generator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.7
sentence-transformers==2.7.0
transformers==4.40.0
huggingface_hub==0.23.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
regex==2023.10.3
beautifulsoup4==4.12.2
requests==2.31.0
aiofiles==23.2.1
torch==2.3.0
